requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# LangGraph and LangChain
langgraph>=0.0.55
langchain-core>=0.1.0
langchain-openai>=0.1.0

# OpenAI
openai>=1.3.0

# Google Calendar API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.108.0

# HTTP client for Slack
httpx>=0.25.0

# Twilio Video
twilio>=8.11.0

# Configuration management
pydantic-settings>=2.0.0

# Audio processing (Step2 voice enhancement)
numpy>=1.24.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
aiohttp>=3.8.0
pyyaml>=6.0.0

# Code quality
ruff>=0.1.0
mypy>=1.7.0

# Production server (optional)
gunicorn>=21.2.0